+++ lib/rimless/configuration.rb
@@ -2,8 +2,46 @@
 
 module Rimless
   # The configuration for the rimless gem.
-  class Configuration
-    include ActiveSupport::Configurable
+  class Configuration < ActiveSupport::OrderedOptions
+    # Track our configurations settings (+Symbol+ keys) and their defaults as
+    # lazy-loaded +Proc+'s values
+    class_attribute :defaults,
+                    instance_reader: true,
+                    instance_writer: false,
+                    instance_predicate: false,
+                    default: {}
+
+    # Create a new +Configuration+ instance with all settings populated with
+    # their respective defaults.
+    #
+    # @param args [Hash{Symbol => Mixed}] additional settings which
+    #   overwrite the defaults
+    # @return [Configuration] the new configuration instance
+    def initialize(**args)
+      super()
+      defaults.each { |key, default| self[key] = instance_exec(&default) }
+      merge!(**args)
+    end
+
+    # A simple DSL method to define new configuration accessors/settings with
+    # their defaults. The defaults can be retrieved with
+    # +Configuration.defaults+ or +Configuration.new.defaults+.
+    #
+    # @param name [Symbol, String] the name of the configuration
+    #   accessor/setting
+    # @param default [Mixed, nil] a non-lazy-loaded static value, serving as a
+    #   default value for the setting
+    # @param block [Proc] when given, the default value will be lazy-loaded
+    #   (result of the Proc)
+    def self.config_accessor(name, default = nil, &block)
+      # Save the given configuration accessor default value
+      defaults[name.to_sym] = block || -> { default }
+
+      # Compile reader/writer methods so we don't have to go through
+      # +ActiveSupport::OrderedOptions#method_missing+.
+      define_method(name) { self[name] }
+      define_method("#{name}=") { |value| self[name] = value }
+    end
 
     # Used to identity this client on the user agent header
     config_accessor(:app_name) { Rimless.local_app_name }
@@ -67,7 +105,7 @@ module Rimless
     #
     # @param val [String, Symbol] the new job queue name
     def consumer_job_queue=(val)
-      config.consumer_job_queue = val.to_sym
+      self[:consumer_job_queue] = val.to_sym
       # Refresh the consumer job queue
       Rimless::ConsumerJob.sidekiq_options(
         queue: Rimless.configuration.consumer_job_queue