+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### next
 
 * Added support for Rails 8.1 (#60)
+* Switched from `ActiveSupport::Configurable` to a custom implementation based
+  on `ActiveSupport::OrderedOptions` (#61)
 
 ### 2.0.0 (28 June 2025)